Add IsValidPosition and update Entity Constructor
Added Entity::IsValidPosition, to check if a position is valid relative to the Environment object given. Each environment object only has 1 grid of a fixed size, so if the entity's position falls outside of that, this will return false. Also updated exception handling in the constructor to utilize this function.