Update DisplayGenerationData() function
The function variable generationNum, which is used to display the current generation value of entity data being displayed, will now update for prior to the start of each new generation. Since each generation's data is displayed starting with the entity with id number 1, a new generation is determined as the point right before the data for entity 1 has its data displayed.