Move output from one class function to another
When an entity moves or food spawns, a message is displayed. Originally, the message would be displayed in the simulation class, which called the movement/spawn function. Now, the message is instead displayed within the movement/spawn funciton itself, assuming the action was completed successfully. This allows for more detailed information to be easily displayed, specifically the coordinates of where the entity moved or where the food was spawned.