Change name and value of offspringNum
Value name is changed in both the Entity class and the EntityData struct. This value now starts at 1 instead of 0. Instead of representing the number of offspring, it now represents the number of unique lineages that have occupied the given slot(which can also be viewed as the number of times an entity has not survived in the slot + 1).